refactor: Extract WASM loading to reusable utility and fix test-wasm.html
Address code review findings:

1. Critical fix: Add locateFile to test-wasm.html for consistent WASM loading
2. Refactor: Extract duplicate WASM loading logic to web/src/utils/wasmLoader.ts
   - Reduces code duplication across all three simulation hooks
   - Provides consistent error handling and path resolution
   - Simplifies future maintenance

Changes:
- Create wasmLoader.ts with loadSopotWasmModule() utility
- Update useRocketSimulation.ts to use utility (simplified from 45 to 10 lines)
- Update useGrid2DSimulation.ts to use utility (simplified from 35 to 7 lines)
- Update useInvertedPendulumSimulation.ts to use utility (simplified from 35 to 7 lines)
- Fix test-wasm.html to include locateFile callback

Benefits:
- DRY principle: ~110 lines of duplicate code eliminated
- Consistent behavior across all simulators
- Single point of maintenance for WASM loading logic
- Easier to add features like retry logic, caching, or integrity checks
