+++ web/src/hooks/useInvertedPendulumSimulation.ts
@@ -1,6 +1,7 @@
 import { useState, useEffect, useRef, useCallback } from 'react';
 import type { SopotModule } from '../types/sopot';
 import type { PendulumState, PendulumVisualizationData } from '../components/InvertedPendulumVisualization';
+import { loadSopotWasmModule } from '../utils/wasmLoader';
 
 /**
  * Interface for the InvertedPendulumSimulator from WASM
@@ -111,36 +112,8 @@ export function useInvertedPendulumSimulation() {
       try {
         console.log('[Pendulum] Loading WASM module...');
 
-        // Use base URL to handle GitHub Pages deployment path
-        const basePath = import.meta.env.BASE_URL || '/';
-        const moduleUrl = `${basePath}sopot.js`;
-
-        console.log(`[Pendulum] Loading from: ${moduleUrl}`);
-
-        // Fetch the module and create a blob URL to bypass Vite's transform
-        const response = await fetch(moduleUrl);
-        if (!response.ok) {
-          throw new Error(`Failed to fetch ${moduleUrl}: ${response.status}`);
-        }
-        const moduleText = await response.text();
-        const blob = new Blob([moduleText], { type: 'application/javascript' });
-        const blobUrl = URL.createObjectURL(blob);
-
-        // @ts-ignore - Dynamic import of WebAssembly
-        const createSopotModule = await import(/* @vite-ignore */ blobUrl);
-        URL.revokeObjectURL(blobUrl);
-
-        if (!mounted) return;
-
-        const module = await createSopotModule.default({
-          locateFile: (path: string) => {
-            // WASM file should be loaded from the public directory
-            if (path.endsWith('.wasm')) {
-              return `${basePath}${path}`;
-            }
-            return path;
-          }
-        });
+        // Load the SOPOT WebAssembly module
+        const module = await loadSopotWasmModule();
 
         if (!mounted) return;
